More packets changed, unique IDs D:

+++ src/pocketmine/network/protocol/AddEntityPacket.php
@@ -48,12 +48,12 @@ class AddEntityPacket extends DataPacket{
 
 	public function encode(){
 		$this->reset();
-		$this->putInt($this->eid);
+		$this->putLong($this->eid);
 		$this->putInt($this->type);
 		$this->putFloat($this->x);
 		$this->putFloat($this->y);
 		$this->putFloat($this->z);
-		$this->putInt($this->did);
+		$this->putLong($this->did);
 		if($this->did > 0){
 			$this->putShort($this->speedX * 8000);
 			$this->putShort($this->speedY * 8000);
